+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Version 4.1.1 (Under development)
 
+### App Center Crashes
+
+* **[Fix]** Fix formatting of stack trace in the `ErrorReport`.
+
 ___
 
 ## Version 4.1.0